Phylogenetic tree of marine fish hosts of Caligus and Lepeophtheirus

A phylogenetic tree (class phylo) of 625 marine fish species that are recorded hosts of sea lice of the genera Caligus and Lepeophtheirus in the cofid copepod-fish interaction database. The tree is pruned from the Fish Tree of Life actinopterygian phylogeny (Rabosky et al. 2018, actinopt_12k_raxml). Tip labels are binomial species names with spaces.

Usage

fish_tree

Format

A phylo object with 625 tips.

Source

https://fishtreeoflife.org/. The underlying data are released under the CC0 1.0 public domain dedication (Dryad, doi:10.5061/dryad.fc71cp4 ). Please cite Rabosky, D. L. et al. (2018) An inverse latitudinal gradient in speciation rate for marine fishes. Nature 559, 392-395. doi:10.1038/s41586-018-0273-1

See also